requirements.txt
dgl
wandb
numpy==1.22.1
pandas
pytorch
rdkit